requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.87
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.0
datasets==2.13.1
decorator==5.1.1
deprecation==2.1.0
dill==0.3.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.96.1
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
img2pdf==0.4.4
imgaug==0.4.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.12.2
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlenlp==2.4.7
paddleocr==2.6.1.2
paddlepaddle==2.4.1
pandas==2.0.3
pdf2docx==0.5.6
pikepdf==8.1.1
Pillow==9.3.0
plotly==5.15.0
premailer==3.10.0
protobuf==3.20.0
psutil==5.9.5
pyarrow==12.0.1
pyclipper==1.3.0.post4
pycryptodome==3.18.0
pydantic==1.10.11
Pygments==2.15.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pypdfium2==4.18.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.13.0
python-engineio==4.5.1
python-multipart==0.0.5
python-socketio==5.8.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.1.1
rarfile==4.0
redis==4.6.0
reflex==0.2.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.4.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.99
seqeval==1.2.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
starlette-admin==0.9.0
tenacity==8.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.13.3
torch==1.13.1
tqdm==4.65.0
transformers==4.25.1
typer==0.4.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.20.0
visualdl==2.5.3
watchdog==2.3.1
watchfiles==0.19.0
websockets==10.4
Werkzeug==2.3.6
xxhash==3.2.0
yarl==1.9.2
